validate create orientation
https://hamoria.atlassian.net/browse/HM-186
