--- middleware/validateOrientationMiddlewar.js
@@ -1,17 +1,0 @@
-import { body, validationResult } from 'express-validator'
-import { BadRequestError } from '../errors/customErrors'
-
-const withValidationErrors = (validateValues) => {
-  return [
-    validateValues,
-    (req, res, next) => {
-      const errors = validationResult(req)
-
-      if (!errors.isEmpty()) {
-        const errorMessages = errors.array().map((error) => error.msg)
-        throw new BadRequestError(errorMessages)
-      }
-      next()
-    },
-  ]
-}
